feat: implement show rewards info for guest
