+++ lib/core/presentation/pages/event/event_detail_page/guest_event_detail_page/views/post_guest_event_detail_view.dart
@@ -4,6 +4,7 @@ import 'package:app/core/presentation/pages/event/event_detail_page/guest_event_
 import 'package:app/core/presentation/pages/event/event_detail_page/guest_event_detail_page/widgets/guest_event_detail_dashboard.dart';
 import 'package:app/core/presentation/pages/event/event_detail_page/guest_event_detail_page/widgets/guest_event_detail_hosts.dart';
 import 'package:app/core/presentation/pages/event/event_detail_page/guest_event_detail_page/widgets/guest_event_detail_photos.dart';
+import 'package:app/core/presentation/pages/event/event_detail_page/guest_event_detail_page/widgets/guest_event_detail_rewards.dart';
 import 'package:app/core/presentation/pages/event/event_detail_page/guest_event_detail_page/widgets/guest_event_location.dart';
 import 'package:app/core/presentation/pages/event/event_detail_page/guest_event_detail_page/widgets/guest_event_poap_offers.dart';
 import 'package:app/core/presentation/pages/event/event_detail_page/guest_event_detail_page/widgets/post_guest_event_animated_app_bar.dart';
@@ -54,6 +55,15 @@ class PostGuestEventDetailView extends StatelessWidget {
                   ),
                   SliverPadding(
                     padding: EdgeInsets.only(top: Spacing.smMedium * 2),
+                  ),
+                  SliverPadding(
+                    padding: EdgeInsets.symmetric(horizontal: Spacing.smMedium),
+                    sliver: SliverToBoxAdapter(
+                      child: GuestEventDetailRewards(event: event),
+                    ),
+                  ),
+                  SliverPadding(
+                    padding: EdgeInsets.only(top: Spacing.smMedium),
                   ),
                   SliverPadding(
                     padding: EdgeInsets.symmetric(horizontal: Spacing.smMedium),
