+++ lib/router/app_router.dart
@@ -167,6 +167,9 @@ final eventDetailRoutes = AutoRoute(
       page: ClaimRewardsRoute.page,
     ),
     AutoRoute(
+      page: GuestEventRewardUsesRoute.page,
+    ),
+    AutoRoute(
       page: EventControlPanelRoute.page,
       children: [
         AutoRoute(
